refactored user and profile services and controllers
