+++ src/Core/Features/IUserService.cs
@@ -1,6 +1,14 @@
+using WordRush.Repository.Models;
+
 namespace WordRush.Core.Features;
 
 public interface IUserService
 {
   Task<bool> GetUserByEmail(string email);
+
+  Task<User?> GetUserProfileByEmail(string email);
+
+  Task<User?> UpdateUserProfile(int id, string nickname, string avatar, string email);
+
+  Task<bool> ChangeUserPassword(int userId, string currentPassword, string newPassword);
 }
