+++ src/Core/WordRush.Core.csproj
@@ -7,6 +7,8 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <Compile Remove="Features\IProfileService.cs" />
+    <Compile Remove="Features\ProfileService.cs" />
     <Compile Remove="GlobalSuppressions.cs" />
   </ItemGroup>
